Moving utils files to the utils folder (keeping the lib folder with handler related functions)
