+++ server/services/zwavejs-ui/utils/convertToGladysDevice.js
@@ -2,8 +2,8 @@ const cleanNames = require('./cleanNames');
 
 const { EXPOSES, PARAMS, COMMANDCLASS, PRODUCTID } = require('../lib/constants');
 const getProperty = require('./getProperty');
-const { refineCategory } = require('../lib/zwaveJSUI.refineCategory');
-const { getDeviceFeatureName } = require('../lib/zwaveJSUI.getDeviceFeatureName');
+const { refineCategory } = require('./refineCategory');
+const { getDeviceFeatureName } = require('./getDeviceFeatureName');
 
 const getDeviceFeatureId = (nodeId, commandClassName, endpoint, propertyName, propertyKeyName, featureName) => {
   const propertyKeyNameClean = cleanNames(propertyKeyName);
